+++ RELEASENOTES.md
@@ -31,3 +31,4 @@
 - Fix changing skill during gameplay to correctly set on next level load
 - Fix setting skill from command line changing the skill level from a save when a new map is loaded
 - Fix issue where players view can be unintentionally changed from mouse input during melt transition
+- Fix UMAPINFO default mapping for secret exit text levels